use user to get projectname in shoot info
